requirements.txt
tqdm
PyQt5
PyMuPDF
Pillow
